An Attempt to Resolve Issues Regarding Climber
The result of series of failed attempts to change modes.

In this version you click on a cub near the first wall to try to change
modes. Long story short it doesnt work. The climber object is complex
and apparently I'm unable to get it to properly activate in game.

At first I tried instantiating it from a prefab (which I put in a
resource folder I later got rid of) and that didn't work. In this
version I tried having the objects on the scene but deactivated and
stored them statically on startup and reactivating them. In both cases
the objects would show up but would for some reason respond to keyboard
input and fall if not touched ( no clinging occured).

Really not sure what to do at the moment. I'm gonna sleep on it and
resume in the morning.