Mixed Bag
Disabled the objects I was trying to manually de and re activate as they
were not achieving our ends and getting in the way.

Transitions from First person controller to wall climber still not
working.

On the bright side though I put a basic (and in some cases somewhat
filled) version of the error node network into each wall. So far only
one of them has an actual error node but all of them do have an end
node. As it stands though w still have to define the behavior on
reaching an end node.